% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCircleLegend.R
\name{addCircleLegend}
\alias{addCircleLegend}
\title{SEACAR addCircleLegend to maps}
\usage{
addCircleLegend(
  map,
  title = "",
  range,
  scaling_fun,
  ...,
  color,
  weight,
  fillColor,
  fillOpacity,
  position = c("topright", "bottomright", "bottomleft", "topleft"),
  data = leaflet::getMapData(map),
  layerId = NULL,
  type = "discrete"
)
}
\value{
A `leaflet`-based function
}
\description{
A customized function to add Circle Legends to static SEACAR Atlas maps. Modified from GitHub user PaulC91:
https://gist.github.com/PaulC91/31c05f84b25975047092f13c2474507a
}
\examples{
library(leaflet)

calc_radius_sav <- function(n){sqrt(n)}

map <- map \%>\%
  addCircleMarkers(data = pts,
                   lat=~Latitude_D, lng=~Longitude_,
                   weight=1, color = "#000000", stroke = TRUE,
                   fillColor = ~color,
                   radius=calc_radius_sav(pts$n_data),
                   fillOpacity=~alpha)

map <- leaflet() \%>\% addTiles() \%>\%
  addCircleLegend(title = "Number of samples",
                  range = c(lns$n_data, pts$n_data),
                  scaling_fun = calc_radius_sav,
                  fillColor = "#b3b3b3",
                  fillOpacity = 0.8,
                  weight = 1,
                  color = "#000000",
                  position = "topright",
                  type = "sav")
}
