% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R
\name{breaks}
\alias{breaks}
\title{Breaks function for Continuous and Discrete WQ Plots}
\usage{
breaks(plot_data, type = "Discrete", ret = "break")
}
\arguments{
\item{plot_data}{A dataframe containing data to be plotted, including Year and YearMonthDec}

\item{type}{"Discrete" or "Continuous"}

\item{ret}{type of return, either "break" to return a sequence of years or "lims" to return min and max year values}
}
\value{
A sequence of breaks or an array of limits
}
\description{
Function to determine x-axis breaks and labels.
}
