% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTrends.R
\name{checkTrends}
\alias{checkTrends}
\title{checkTrends function for parsing statistical results}
\usage{
checkTrends(p, Slope, SufficientData)
}
\arguments{
\item{p}{p-value}

\item{Slope}{LME Slope value}

\item{SufficientData}{boolean value (TRUE/FALSE)}
}
\value{
Character-based statement denoting trend
}
\description{
Function to create text for stats results.
}
