% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_shape.R
\name{find_shape}
\alias{find_shape}
\title{SEACAR find_shape function to locate individual Managed Area boundary}
\usage{
find_shape(rcp, ma)
}
\arguments{
\item{rcp}{sf dataframe corresponding to the output from SEACAR::GeoData$rcp}

\item{ma}{Managed Area Name (full name)}
}
\value{
A single-line sf dataframe
}
\description{
SEACAR find_shape function to locate individual Managed Area boundary
}
\examples{
ma_shape <- find_shape(rcp = SEACAR::GeoData$rcp, ma = "Estero Bay Aquatic Preserve")
}
