% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_coral_description.R
\name{generate_coral_description}
\alias{generate_coral_description}
\title{SEACAR generate_coral_description function to generate table descriptions for Coral habitat for a single ManagedArea, to be used within `generate_description()`.}
\usage{
generate_coral_description(data)
}
\arguments{
\item{data}{a dataframe with statistical analysis results for Coral habitat.}
}
\value{
A text-based descriptive statement wrapped in HTML paragraph tags.
}
\description{
SEACAR generate_coral_description function to generate table descriptions for Coral habitat for a single ManagedArea, to be used within `generate_description()`.
}
