% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cw_description.R
\name{generate_cw_description}
\alias{generate_cw_description}
\title{SEACAR generate_cw_description function to generate table descriptions for Coastal Wetlands habitat for a single ManagedArea, to be used within `generate_description()`.}
\usage{
generate_cw_description(data)
}
\arguments{
\item{data}{a dataframe with statistical analysis results for Coastal Wetland habitat.}
}
\value{
A text-based descriptive statement wrapped in HTML paragraph tags.
}
\description{
SEACAR generate_cw_description function to generate table descriptions for Coastal Wetlands habitat for a single ManagedArea, to be used within `generate_description()`.
}
