% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_nekton_description.R
\name{generate_nekton_description}
\alias{generate_nekton_description}
\title{SEACAR generate_nekton_description function to generate table descriptions for Nekton habitat for a single ManagedArea, to be used within `generate_description()`.}
\usage{
generate_nekton_description(data)
}
\arguments{
\item{data}{a dataframe with statistical analysis results for Nekton habitat.}
}
\value{
A text-based descriptive statement wrapped in HTML paragraph tags.
}
\description{
SEACAR generate_nekton_description function to generate table descriptions for Nekton habitat for a single ManagedArea, to be used within `generate_description()`.
}
