% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sav_description.R
\name{generate_sav_description}
\alias{generate_sav_description}
\title{SEACAR generate_nekton_description function to generate table descriptions for SAV habitat for a single ManagedArea, to be used within `generate_description()`.}
\usage{
generate_sav_description(data)
}
\arguments{
\item{data}{a dataframe with statistical analysis results for SAV habitat.}
}
\value{
A text-based descriptive statement wrapped in HTML paragraph tags.
}
\description{
SEACAR generate_nekton_description function to generate table descriptions for SAV habitat for a single ManagedArea, to be used within `generate_description()`.
}
