% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_wc_description.R
\name{generate_wc_description}
\alias{generate_wc_description}
\title{SEACAR generate_wc_description function to generate table descriptions for WC Discrete and Continuous habitat for a single ManagedArea, to be used within `generate_description()`.}
\usage{
generate_wc_description(input_df)
}
\arguments{
\item{input_df}{a dataframe with statistical analysis results for WC Discrete and Continuous.}
}
\value{
A text-based descriptive statement wrapped in HTML paragraph tags.
}
\description{
SEACAR generate_wc_description function to generate table descriptions for WC Discrete and Continuous habitat for a single ManagedArea, to be used within `generate_description()`.
}
