% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoData.R
\docType{data}
\name{GeoData}
\alias{GeoData}
\title{GeoData}
\format{
A list with 3 spatial (sf) dataframes:
\describe{
  \item{pointLocations}{SEACAR sample location values (single point coordinates - POINT `geometry`)}
  \item{lineLocations}{SEACAR sample location values (transect-based coordinates - MULTILINESTRING `geometry`)}
  \item{RCP Boundaries}{ORCP Managed Area boundaries}
}
}
\source{
Created by Florida SEACAR
}
\usage{
GeoData
}
\description{
Three spatial sf-DataFrames containing sample locations and RCP boundaries
}
\keyword{datasets}
