% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shape_coordinates.R
\name{get_shape_coordinates}
\alias{get_shape_coordinates}
\title{SEACAR get_shape_coordinates function to provide coordinate min and max to allow for accurate setting of viewport on Leaflet maps}
\usage{
get_shape_coordinates(ma_shape)
}
\arguments{
\item{ma_shape}{a single-line sf dataframe (output from SEACAR::find_shape())}
}
\value{
A data.frame with 4 values corresponding to min and max coordinate values: xmin, ymin, xmax, ymax
}
\description{
SEACAR get_shape_coordinates function to provide coordinate min and max to allow for accurate setting of viewport on Leaflet maps
}
\examples{
shape_coordinates <- get_shape_coordinates(ma_shape)
map <- leaflet() \%>\% addTiles() \%>\%
  fitBounds(lng1=shape_coordinates$xmin,
            lat1=shape_coordinates$ymin,
            lng2=shape_coordinates$xmax,
            lat2=shape_coordinates$ymax)
}
