% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagedAreas.R
\docType{data}
\name{ManagedAreas}
\alias{ManagedAreas}
\title{ManagedAreas}
\format{
A data frame with columns:
\describe{
  \item{AreaID}{numeric value assigned to each Managed Area}
  \item{ManagedAreaName}{full name associated with a given Managed Area}
  \item{ShortName}{shortened name (e.g., "AP", "NERR")}
  \item{Region}{FDEP regional association}
  \item{Abbreviation}{abbreviated form of ManagedArea}
}
}
\source{
Created by Florida SEACAR
}
\usage{
ManagedAreas
}
\description{
A reference of all available RCP ManagedAreas available in SEACAR.
}
\keyword{datasets}
