% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEACAR-package.R
\docType{package}
\name{SEACAR}
\alias{SEACAR-package}
\alias{SEACAR}
\title{SEACAR: Statewide Ecosystem Assessment of Coastal and Aquatic Resources}
\description{
A package for data analysis, visualization, and reporting of water quality and habitat monitoring data
from the Florida SEACAR program. Provides standardized datasets, analysis functions, themes, and tools.
}
\keyword{internal}
