% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEACAR_palette1.R
\name{seacar_palette1}
\alias{seacar_palette1}
\title{SEACAR Palette1}
\usage{
seacar_palette1
}
\value{
A vector of hex colors, primarily used for species-based plots (i.e., SAV Frequency of Occurrence Barplots)
}
\description{
SEACAR Palette1
}
