% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEACAR_palette1_SAV.R
\name{seacar_palette1_sav}
\alias{seacar_palette1_sav}
\title{SEACAR Palette1 (SAV species associations)}
\usage{
seacar_palette1_sav
}
\value{
A named vector of hex colors, containing SAV species associations
}
\description{
SEACAR Palette1 (SAV species associations)
}
