% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEACAR_plot_theme.R
\name{SEACAR_plot_theme}
\alias{SEACAR_plot_theme}
\title{SEACAR ggplot2 theme}
\usage{
SEACAR_plot_theme()
}
\value{
A `ggplot2` theme object
}
\description{
A custom `ggplot2` theme conforming to Florida SEACAR visualization standards.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(mpg, wt)) +
  geom_point() +
  SEACAR_plot_theme()
}
