% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subchunkify.R
\name{subchunkify}
\alias{subchunkify}
\title{Subchunkify function for RMarkdown}
\usage{
subchunkify(g, fig_height = 9, fig_width = 10, fig_cap)
}
\arguments{
\item{g}{Knitr-based input to be parsed}

\item{fig_height}{Figure height (inches)}

\item{fig_width}{Figure width (inches)}

\item{fig_cap}{figure caption text}
}
\value{
A deparsed knitr subchunk
}
\description{
Function to allow dynamic figure resizing and proper caption implementation. Developed by Michael JW, via Stackoverflow user `jzadra`.
}
