% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_description.R
\name{generate_description}
\alias{generate_description}
\title{SEACAR generate_description function to generate table descriptions for various habitats.}
\usage{
generate_description(data, habitat)
}
\arguments{
\item{data}{a dataframe with statistical analysis results for various habitats.}

\item{habitat}{a string containing the habitat to generate descriptions for. Options: "Coral", "CW", "Nekton", "Oyster", "SAV", "WC"}
}
\value{
A text-based descriptive statement wrapped in HTML paragraph tags.
}
\description{
SEACAR generate_description function to generate table descriptions for various habitats.
}
