% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_oyster_description.R
\name{generate_oyster_description}
\alias{generate_oyster_description}
\title{SEACAR generate_osyter_description function to generate table descriptions for Oyster habitat for a single ManagedArea, to be used within `generate_description()`.}
\usage{
generate_oyster_description(data)
}
\arguments{
\item{data}{a dataframe with statistical analysis results for Oyster habitat.}
}
\value{
A text-based descriptive statement wrapped in HTML paragraph tags.
}
\description{
SEACAR generate_osyter_description function to generate table descriptions for Oyster habitat for a single ManagedArea, to be used within `generate_description()`.
}
