% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEACAR_palette1_SAV.R
\docType{data}
\name{seacar_palette1_sav}
\alias{seacar_palette1_sav}
\title{SEACAR Palette1 (SAV species associations)}
\format{
An object of class \code{character} of length 13.
}
\usage{
seacar_palette1_sav
}
\value{
A named vector of hex colors, containing SAV species associations
}
\description{
SEACAR Palette1 (SAV species associations)
}
\keyword{datasets}
