% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEACAR_palette2.R
\docType{data}
\name{seacar_palette2}
\alias{seacar_palette2}
\title{SEACAR Palette2}
\format{
An object of class \code{character} of length 15.
}
\usage{
seacar_palette2
}
\value{
A 15-color palette to be used on maps and other plots where a larger number of colors are needed, usually when distinguishing between programs
}
\description{
SEACAR Palette2
}
\keyword{datasets}
